Fix: network: remove ifcfg-* files when interface is detached
When the interface is detached, the sctipt can't get the original device
name needed to delete ifcfg-* file.

If the MAC is not found in the environment, search deleted dir for
ETHx_MAC definitions.

If the device is not found in `ip list` by the MAC address (because it
is already detached), search in /etc/sysconfig/network-scripts/ifcfg-*
for HWADDR definitions.